Add Dead Event and Anim

+++ Assets/Scripts/Game Saving/CharacterSaveData.cs
@@ -23,7 +23,7 @@ namespace Game_Saving
         public float zPosition;
         
         [Header("Resources")]
-        public float currentHealth;
+        public int currentHealth;
         public float currentStamina;
         
         [Header("Stats")]
